Set input_delay_ns in the spi_device_interface_config_t
This tweaks the internal computation for half duplex mode work
arounds of when to capture / how much to shift.

Note that we don't set it for Full Duplex mode. In theory
with the delay we have, it shouldn't work above 20 MHz according
to the SDK ... but it does, so leave it at 0 there since in the
full duplex case this value is only used for checks and doesn't
actually influence the hw config.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>